+++ tools/schematics/evolution/schema.ts
@@ -10,6 +10,8 @@ export type EvolutionName =
 export interface EvolutionOptions {
   readonly name: EvolutionName;
   readonly preview?: boolean;
+  readonly translocoLanguages?: string;
+  readonly translocoDefaultLanguage?: string;
   readonly storeScope?: 'feature' | 'root';
   readonly featureName?: string;
   readonly featureComponent?: 'existing' | 'create';